+++ spec/rubocop/cops/case_indentation_spec.rb
@@ -4,8 +4,8 @@ require 'spec_helper'
 
 module Rubocop
   module Cop
-    describe Indentation do
-      let(:ind) { Indentation.new }
+    describe CaseIndentation do
+      let(:ind) { CaseIndentation.new }
 
       it "registers an offence for a when clause that's deeper than case" do
         source = ['case a',